Total layout redesign.
App bar and sidebar is removed. Bottom nav that adapts to mobile and desktop is now used. Students data cache is removed. It causes homepage to continuously re-render.

This is done to reclaim much unused space in our preparation to move to our next phase of development, desktop optimized layout and reporting.

For future caching effort, it might be better to rely on react-query. Why? because it's smaller than axios, simpler to implement compared to doing it manually, supports the upcoming react suspense, and also supported by a lot of other people.