Fix bugs cause by react-query. Scraps PageAnalyze.
PageAnalyze is scrapped for a simpler alternative solution. Fix dialog not scrollable.

Scrollable dialog seems to be a bad idea, so are any complex ui inside a dialog. The fact that dialog has a fixed position causes some bug that are hard to deal with, especially on iOS. Dialog should be kept in minimal use. Only used for simpler stuff.