Update snapshots. Use gatsby-plugin-sentry.
Loading sentry through segment is not effective. We delay segment loading until user navigates around the website, which means that when they first land, sentry won't be loaded. This is problematic, since a lot of errors happens directly when user finish loading the page.

To avoid this problem, we now bundle sentry directly inside the main js file. The bigger bundle size is a bit concerning tho, better revisit this again in the future.