Fix test errors.
Password reset test was failing because token that was being returned by authStores.InsertNewToken was different from token that were saved into the DB.

School's test was failing because now we're saving *auth.SessionData to context, instead of the *postgres.Session. And the test hasn't been updated to use the new auth.SessionData.