Use RFC 8061 date string for get plans API.
Just using YYYY-MM-DD actually aren't enough. Since the server saves the data using UTC, when a client from Indonesia send a get plan request using the date 2020-12-02, the server will returns all plans for 2020-12-02 UTC, not UTC+7 (aka WIB timezone).

Because of this, from now on all date query will require a timezone.

Using RFC 8061 means that we wont need to do conversions manually (dayjs and browser Intl API handles it on client, postgres handles it on backend).

Another approach to this might be to use unix epoch on API request, but this has some drawbacks:
1. Clients will need to remember to always convert to UTC. This will make clients more error prone. Client codes usually also change a lot more frequently than backend codes, which exacerbate this issue.
2. Its not human readable.