ref: Add FileStorage interface to schoolStore.
In order to shift the responsibility of wrangling data saving and deleting out of the handlers, FileStorage interface is now added to the schoolStore directly.

An example of the data flow will look like this:
1. Handler for post `/schools/{schoolId}/files` receives a request containing a file.
2. It passes the file, the name of the file and the schoolId to the schoolStore.
3. SchoolStore will then try to save it to minio using FileStorage interface.
4. On successful file save, it will then save the record to db.

Here instead of the handler having to pass save the file, check if it is successful and then save the data to the db, we can just let schoolStore deal with those and let the handler deal with the data mapping and request/response validation.