fix: Fix race condition on useCacheNewStudentFormData
A small fascinating bug, sometimes when we run on slow devices, new student form's cache might not get reset after a user finishes submitting the form. This is caused by a few issues:
- race condition when updating cache
- cache update not debounced
- different cache update order between resetting cache and updating cache with new form data.

The race condition is usually fine when devices are fast. But on extremely slow devices, sometimes the cache update call can pile up. This is made worse by useCacheNewStudentFormData that is not debounced. Which means that cache update call were happening on each keystroke and make the chances that cache update might pile up higher. This is usually fine because the cache would still be updated in the order of the call, first in first serve.

We have 2 different cache, student profile image in indexedDB and student json data on local storage. The ordering of update between this 2 cache is what  finally trigger the bug. This ordering usually doesn't matter when cache update doesn't pile up, but when it finally does, different ordering of cache update will cause a race condition which will make the latest update got executed before some earlier one finishes. This means that earlier update might override the data of later ones.

To prevent this, now:
- useCacheNewStudentFormData is debounced by 100ms
- all new student form's cache update uses the same function so that the all update follows the same ordering.